feat: initial support for passing args to `terraform plan`
No special handling of shell arguments is made here. The action
configuration needs to account for shell escaping for now.

This is not really ideal, but given how multiple directories were added in
f39c110 (with list of directories separated by spaces) and a matching array
with the tf versions, I'm not really sure of a better way.

Perhaps eventually a new major release where we change how all the parameters
are passed could sort it out so that instead we are explicit about the tuples
of directory+tfversion+args instead of trying to match them across three
arrays.

Anyway, for our use case, this change allows us to reduce a few hundred lines
of terraform plan output into an empty one since we can pass in the version
numbers instead of defaulting to `latest` during pull request pr checks.